Use uint8_t for TopologyLocation size
Reduces size of a TopologyLocation from 16 bytes to 4

+++ include/geos/geomgraph/TopologyLocation.h
@@ -132,7 +132,7 @@ public:
 private:
 
     std::array<geom::Location, 3> location;
-    std::size_t locationSize;
+    std::uint8_t locationSize;
 };
 
 std::ostream& operator<< (std::ostream&, const TopologyLocation&);
